# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: VIB DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/annual-plans/{plan_id}/projects:
    get:
      summary: List projects linked to plan (FR-021)
      operationId: listPlanProjects
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of linked projects

    post:
      summary: Link project to plan (FR-021, BR-009)
      operationId: linkProjectToPlan
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id]
              properties:
                project_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Link created
        "409":
          description: PLAN_NOT_ACTIVE — plan must be active (BR-009)
  /api/v1/annual-plans/{plan_id}/projects/{project_id}:
    delete:
      summary: Unlink project from plan (FR-021)
      operationId: unlinkProjectFromPlan
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unlinked (soft — sets unlinked_at)

  # =========================================================================
  # Module 2 — Project Objects (PPG :8001)
  # =========================================================================
...